Tratamento de exceção

+++ src/main/java/com/devlpjruan/ecommercproject/services/ProductService.java
@@ -1,14 +1,19 @@
 package com.devlpjruan.ecommercproject.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devlpjruan.ecommercproject.dto.ProductDto;
 import com.devlpjruan.ecommercproject.entities.Product;
 import com.devlpjruan.ecommercproject.repository.ProductRepository;
+import com.devlpjruan.ecommercproject.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -18,12 +23,13 @@ public class ProductService {
 
 	@Transactional(readOnly = true)
 	public ProductDto findById(Long id) {
-		Product product = productRepository.findById(id).get();
+		Product product = productRepository.findById(id).orElseThrow(
+				() -> new ResourceNotFoundException("Recurso não encontrado"));
 		return new ProductDto(product);
 	}
 	
 	@Transactional(readOnly = true)
-	public Page<Object> findAll(Pageable pageable) {
+	public Page<ProductDto> findAll(Pageable pageable) {
 		Page<Product> result = productRepository.findAll(pageable);
 		return result.map(x -> new ProductDto(x));
 	}
@@ -38,22 +44,36 @@ public class ProductService {
 	}
 	@Transactional
 	public ProductDto updateProduct(Long id, ProductDto dto) {
+		try {
 		Product entity = productRepository.getReferenceById(id);
 		DtoToEntity(dto, entity);
 		entity = productRepository.save(entity);
 		return new ProductDto(entity);
+		
+		}catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 	
 	private void DtoToEntity(ProductDto dto, Product entity){
 		entity.setName(dto.getName());
 		entity.setDescription(dto.getDescription());
 		entity.setPrice(dto.getPrice());
-		entity.setImgurl(dto.getImgurl());
+		entity.setImgUrl(dto.getImgurl());
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
+		if(!productRepository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+			
+		}try {
 		 productRepository.deleteById(id);
+		 
+		}catch(DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+			
+		}
 	}
 	
 }
